+++ app/models/__init__.py
@@ -36,6 +36,14 @@ from app.models.featured_course import FeaturedCourse, FeaturedPage, FeaturedSta
 from app.models.follow import Follow, FollowStatus
 from app.models.forum import CourseForum, CourseForumAnswer
 from app.models.learning import CourseLearning
+from app.models.meeting import (
+    DayLabel,
+    Meeting,
+    MeetingTime,
+    ReserveMeeting,
+    ReserveMeetingType,
+    ReserveStatus,
+)
 from app.models.newsletter import Newsletter
 from app.models.noticeboard import CourseNoticeboard, NoticeboardColor
 from app.models.notification import (
@@ -98,6 +106,12 @@ __all__ = [
     "TextLesson",
     "CourseSession",
     "CourseLearning",
+    "Meeting",
+    "MeetingTime",
+    "ReserveMeeting",
+    "DayLabel",
+    "ReserveStatus",
+    "ReserveMeetingType",
     "Quiz",
     "QuizQuestion",
     "QuizQuestionAnswer",